ch04/requirements.txt
beautifulsoup4==4.14.3
parse==1.20.2
delorean==1.0.0
Pillow==12.0.0
xmltodict==1.0.2
pytesseract==0.3.13
easyocr==1.7.2
PyPDF2==3.0.1
python-docx==1.2.0
charset-normalizer

